Use @blueprint.record_once to modify flask-security emails
...instead of fudging factory.py
