+++ hepdata/factory.py
@@ -53,29 +53,12 @@ create_api = create_app_factory(
 )
 
 
-def create_app(*args, **kwargs):
-    factory = create_app_factory(
-        'hepdata',
-        config_loader=conf_loader,
-        extension_entry_points=['invenio_base.apps'],
-        blueprint_entry_points=['invenio_base.blueprints'],
-        wsgi_factory=create_wsgi_factory({'/api': create_api}),
-        instance_path=instance_path,
-        static_folder=static_folder,
-    )
-    app = factory()
-
-    # Customise flask-security emails
-    security = app.extensions['security']
-
-    @security.send_mail_task
-    def send_hepdata_mail(msg):
-        from hepdata.modules.email.utils import send_flask_message_email
-        send_flask_message_email(msg)
-
-    @security.mail_context_processor
-    def security_mail_processor():
-        site_url = app.config.get('SITE_URL', 'https://www.hepdata.net')
-        return dict(site_url=site_url)
-
-    return app
+create_app = create_app_factory(
+    'hepdata',
+    config_loader=conf_loader,
+    extension_entry_points=['invenio_base.apps'],
+    blueprint_entry_points=['invenio_base.blueprints'],
+    wsgi_factory=create_wsgi_factory({'/api': create_api}),
+    instance_path=instance_path,
+    static_folder=static_folder,
+)
